requirements.txt
Flask
Flask-Cors
# segmentation_models_pytorch
opencv-python
pillow
torchvision